+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log
 
+## [Unreleased]
+
+### Changed
+
+- concord-server: sanitize `\u0000` in strings when inserting JSONB
+data (e.g. process events). 
+
+
+
 ## [1.36.1] - 2019-12-11
 
 ### Changed